FROMLIST: docs: gunyah: Introduce Gunyah Hypervisor
Gunyah is an open-source Type-1 hypervisor developed by Qualcomm. It
does not depend on any lower-privileged OS/kernel code for its core
functionality. This increases its security and can support a smaller
trusted computing based when compared to Type-2 hypervisors.

Add documentation describing the Gunyah hypervisor and the main
components of the Gunyah hypervisor which are of interest to Linux
virtualization development.

Change-Id: Ib8930f467b749ae2adc2d63cbbe0f02ab82f1eba
Reviewed-by: Bagas Sanjaya <bagasdotme@gmail.com>
Signed-off-by: Elliot Berman <quic_eberman@quicinc.com>
Reviewed-by: Alex Elder <elder@linaro.org>
Link: https://lore.kernel.org/lkml/20230613172054.3959700-2-quic_eberman@quicinc.com/
Patch-mainline: linux-arm-kernel @ 13/06/23, 17:20
Signed-off-by: Sayan Dey <quic_sayand@quicinc.com>