FROMLIST: gunyah: rsc_mgr: Add RPC for sharing memory
Gunyah resource manager provides API to manipulate stage 2 page tables.
Manipulations are represented as a memory parcel. Memory parcels
describe a list of memory regions (intermediate physical address--IPA--and
size), a list of new permissions for VMs, and the memory type (DDR or
MMIO). Each memory parcel is uniquely identified by a handle allocated by
Gunyah. There are a few types of memory parcel sharing which Gunyah
supports:

 - Sharing: the guest and host VM both have access
 - Lending: only the guest has access; host VM loses access
 - Donating: Permanently lent (not reclaimed even if guest shuts down)

Memory parcels that have been shared or lent can be reclaimed by the
host via an additional call. The reclaim operation restores the original
access the host VM had to the memory parcel and removes the access to
other VM.

One point to note that memory parcels don't describe where in the guest
VM the memory parcel should reside. The guest VM must accept the memory
parcel either explicitly via a "gh_rm_mem_accept" call (not introduced
here) or be configured to accept it automatically at boot. When the guest
VM accepts the memory parcel, it also mentions the IPA it wants to place
memory parcel. Although the region might be discontiguous on the host,
the memory parcel is place contiguously in the guest memory at the
specified IPA.

Change-Id: I99264b7ea35e7ef7d01e96e5bc42c5e1f29513c8
Reviewed-by: Alex Elder <elder@linaro.org>
Reviewed-by: Srinivas Kandagatla <srinivas.kandagatla@linaro.org>
Co-developed-by: Prakruthi Deepak Heragu <quic_pheragu@quicinc.com>
Signed-off-by: Prakruthi Deepak Heragu <quic_pheragu@quicinc.com>
Signed-off-by: Elliot Berman <quic_eberman@quicinc.com>
Link: https://lore.kernel.org/lkml/20230613172054.3959700-11-quic_eberman@quicinc.com/
Patch-mainline: linux-arm-kernel @ 13/06/23, 17:20
Signed-off-by: Sayan Dey <quic_sayand@quicinc.com>