FROMLIST: gunyah: vm_mgr: Add framework for VM Functions
Introduce a framework for Gunyah userspace to install VM functions. VM
functions are optional interfaces to the virtual machine. vCPUs,
ioeventfs, and irqfds are examples of such VM functions and are
implemented in subsequent patches.

A generic framework is implemented instead of individual ioctls to
create vCPUs, irqfds, etc., in order to simplify the VM manager core
implementation and allow dynamic loading of VM function modules.

Change-Id: Idcafb4a410cd2157c98d31c2a093e2604754df45
Signed-off-by: Elliot Berman <quic_eberman@quicinc.com>
Reviewed-by: Alex Elder <elder@linaro.org>
Link: https://lore.kernel.org/lkml/20230613172054.3959700-19-quic_eberman@quicinc.com/
Patch-mainline: linux-arm-kernel @ 13/06/23, 17:20
Signed-off-by: Sayan Dey <quic_sayand@quicinc.com>