FROMLIST: virt: gunyah: Add irqfd interface
Enable support for creating irqfds which can raise an interrupt on a
Gunyah virtual machine. irqfds are exposed to userspace as a Gunyah VM
function with the name "irqfd". If the VM devicetree is not configured
to create a doorbell with the corresponding label, userspace will still
be able to assert the eventfd but no interrupt will be raised on the
guest.

Change-Id: I4495ccab0eeb0390b3740577f2e4038082c6690a
Acked-by: Alex Elder <elder@linaro.org>
Co-developed-by: Prakruthi Deepak Heragu <quic_pheragu@quicinc.com>
Signed-off-by: Prakruthi Deepak Heragu <quic_pheragu@quicinc.com>
Signed-off-by: Elliot Berman <quic_eberman@quicinc.com>
Reviewed-by: Alex Elder <elder@linaro.org>
Link: https://lore.kernel.org/lkml/20230613172054.3959700-24-quic_eberman@quicinc.com/
Patch-mainline: linux-arm-kernel @ 13/06/23, 17:20
Signed-off-by: Sayan Dey <quic_sayand@quicinc.com>