FROMLIST: ice, ufs, mmc: use blk_crypto_key for program_key
The program key ops in the storage controller does not
pass on the blk crypto key structure to ice, this is okay
when wrapped keys are not supported and keys are standard
AES XTS sizes. However, wrapped keyblobs can be of any size
and in preparation for that, modify the ICE and storage
controller APIs to accept blk_crypto_key.

Signed-off-by: Gaurav Kashyap <quic_gaurkash@quicinc.com>
Link: https://lore.kernel.org/linux-arm-msm/20231122053817.3401748-1-quic_gaurkash@quicinc.com/.

Change-Id: Ie4c32646e6dc3b9c4ea3b7426b66902d920b2ab1
Signed-off-by: Seshu Madhavi Puppala <quic_spuppala@quicinc.com>