FROMLIST: input: pm8xxx-vib: refactor to easily support new SPMI vibrator
Currently, all vibrator control register addresses are hard coded,
including the base address and the offset, it's not flexible to support
new SPMI vibrator module which is usually included in different PMICs
with different base address. Refactor this by defining register offset
with HW type combination, and register base address which is defined
in 'reg' property is added for SPMI vibrators.

Change-Id: I68f4402f1c55ed7eb945088cf5590d0e15b583b7
Link: https://lore.kernel.org/all/20231108-pm8xxx-vibrator-v7-1-632c731d25a8@quicinc.com/
Signed-off-by: Fenglin Wu <quic_fenglinw@quicinc.com>