PENDING: iris: introduce packetization layer for creating HFI packets
Host firmware interface (HFI) is well defined set of interfaces
for communication between host driver and firmware.
The command and responses are exchanged in form of packets.
One or multiple packets are grouped under packet header.
Each packet has packet type which describes the specific HFI
and payload which holds the corresponding value for that HFI.

Sys_init is the first packets sent to firmware, which initializes
the firmware. Sys_image_version packet is to get the firmware
version string.

Change-Id: I482b657e34d39060bac8fca68771541161c382b5
Signed-off-by: Dikshita Agarwal <quic_dikshita@quicinc.com>