BACKPORT: UPSTREAM: OPP: Extend support for the opp-level beyond required-opps
At this point the level (performance state) for an OPP is currently limited
to be requested for a device that is attached to a PM domain.  Moreover,
the device needs to have the so called required-opps assigned to it, which
are based upon OPP tables being described in DT.

To extend the support beyond required-opps and DT, let's enable the level
to be set for all OPPs. More precisely, if the requested OPP has a valid
level let's try to request it through the device's optional PM domain, via
calling dev_pm_domain_set_performance_state().

Change-Id: Icd49b239320b9d1619a7c81564bdd5f615eb820c
Signed-off-by: Ulf Hansson <ulf.hansson@linaro.org>
Git-commit: 0025ff64ffcf6bd6ece5484e7818401f77bf115f
Git-repo: git://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git
[ Viresh: Handle NULL opp in _set_opp_level() ]
Signed-off-by: Viresh Kumar <viresh.kumar@linaro.org>
Signed-off-by: Nikunj Kela <quic_nkela@quicinc.com>