UPSTREAM: media: qcom: camss: Pass CAMSS subdev callbacks via resource ops pointer
It is possible to pass all of the CAMSS subdevice internal operations
pointers from the controlling resources structure with an additional
pointer added to the resources structure.

This allows for the removal of most of the probe-time control structures.

Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>
Signed-off-by: Hans Verkuil <hverkuil-cisco@xs4all.nl>
Git-commit: c23c7998161595b19ee1162da09609c31b5b4393
Git-repo: git://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git
Signed-off-by: Suresh Vankadara <quic_svankada@quicinc.com>