FROMGIT: misc: fastrpc: Free DMA handles for RPC calls with no arguments
The FDs for DMA handles to be freed is updated in fdlist by DSP over
a remote call. This holds true even for remote calls with no
arguments. To handle this, get_args and put_args are needed to
be called for remote calls with no arguments also as fdlist
is allocated in get_args and FDs updated in fdlist is freed
in put_args.

Signed-off-by: Srinivas Kandagatla <srinivas.kandagatla@linaro.org>
Git-commit: 3f01aaeb9fd387108c04c4803b949b8cd84d43ff
Git-repo: git://git.kernel.org/pub/scm/linux/kernel/git/srini/fastrpc.git
Change-Id: Id8abda189b99c3432b3010d4eca3dc5a6a68a905
Signed-off-by: Ekansh Gupta <quic_ekangupt@quicinc.com>