PENDING: iris: register video device to platform driver
Iris is a multi pipe based video acceleration hardware
block that offloads video stream encoding and decoding
from the application processor (AP). It supports H.264
and H.265 encoding and decoding, as well as VP9 decoding.
The AP communicates with hardware through a well defined
protocol which provides fine-grained and asynchronous
control over individual hardware features.

This patch introduces basic probe and remove functions.
It handles setting up a video device as well as registering
it with the V4L2 subsystem.

Change-Id: I36b81845aa462c8cf919371e9e965e9198b47699
Signed-off-by: Dikshita Agarwal <quic_dikshita@quicinc.com>