FROMLIST: usb: dwc3: Override end of dwc3 memory resource
In the case that the dwc3 core driver is instantiated from the same
memory resource information as the glue driver, the dwc_res memory
region will overlap with the memory region already mapped by the glue.

As the DWC3 core driver already does math on the passed memory region to
exclude the XHCI region, also adjust the end address, to avoid having to
pass an adjusted region from the glue explicitly.

Link: https://lore.kernel.org/all/20231016-dwc3-refactor-v1-5-ab4a84165470@quicinc.com/ .

Change-Id: Id303516131943e74da9f3e1d54ad50eb441b131d
Signed-off-by: Bjorn Andersson <quic_bjorande@quicinc.com>
Acked-by: Konrad Dybcio <konrad.dybcio@linaro.org>
Signed-off-by: Krishna Kurapati <quic_kriskura@quicinc.com>