FROMLIST: usb: dwc3: qcom: Add dwc3 core reference in driver state
In the coming changes the Qualcomm DWC3 glue will be able to either
manage the DWC3 core as a child platform_device, or directly instantiate
it within its own context.

Introduce a reference to the dwc3 core state and make the driver
reference the dwc3 core either the child device or this new reference.

As the new member isn't assigned, and qcom->dwc_dev is assigned in all
current cases, the change should have no functional impact.

Link: https://lore.kernel.org/all/20231016-dwc3-refactor-v1-6-ab4a84165470@quicinc.com/ .

Change-Id: I987fea919014d003057b03ca324b7b5dff314074
Signed-off-by: Bjorn Andersson <quic_bjorande@quicinc.com>
Signed-off-by: Krishna Kurapati <quic_kriskura@quicinc.com>