FROMLIST: usb: dwc3: qcom: Enable autosuspend for host mode
When in host mode, enable autosuspend for xhci and root hubs.

a) Register a vendor call back to get information of successful role
change by core. When xhci is enumerated successfully, configure it to
use_autosuspend. The decision of whether or not to do runtime_allow for
xhci node is left to userspace:
(echo auto > */xhci-auto/power/control).

b) Register to usb-core notifications in set_mode vendor callback to
identify when root hubs are being created. Configure them to
use_autosuspend accordingly.

c) Configure any connected device to use_autosuspend. In general for
mobile use cases, autosuspend is enabled and wakeup is enabled only
for hubs and audio devices. So let userspace choose to configure
autosuspend_delay and wakeup capability of connected devices.

Link: https://lore.kernel.org/all/20231017131851.8299-4-quic_kriskura@quicinc.com/ .

Change-Id: I562ce6b18594122c6bc1a6db554376f53bf8d740
Signed-off-by: Krishna Kurapati <quic_kriskura@quicinc.com>