FROMLIST: usb: dwc3: core: Skip set_mode notification if cable is disconnected
In device mode use cases, the following sequence of actions are observed:

1. Cable disconnect happens and clears qscratch HS_PHY_CTRL_REG properly
2. Disconnect event is generated and "connected" flag turns false.
3. Then the setmode notification from core goes to glue
4. Glue will set back the qscratch HS_PHY_CTRL_REG bits again.

At this point, since the cable is removed, setting qscratch bits shouldn't
affect anything. But it is observed that after setting this bits, the
controller generated Event-0x101 and Event-0x30601 (bus reset and suspend)
in order. In bus reset, we set back the "connected" flag and this blocks
suspend again.

So send set_mode call only if the cable is connected, else skip it.

Link: https://lore.kernel.org/all/20231017131851.8299-8-quic_kriskura@quicinc.com/ .

Change-Id: Idafa8dbb93faf834b24ec8fc5483b4ff53dec766
Signed-off-by: Krishna Kurapati <quic_kriskura@quicinc.com>