QCLINUX: arm64: defconfig: qcom: Enable HARDENED_USERCOPY
This option checks for wrong memory regions when copying memory to/from
the kernel (via copy_to_user() and copy_from_user() functions) by
rejecting memory ranges that are larger than the specified heap object,
span multiple separately allocated pages, are not on the process stack,
or are part of the kernel text. This prevents entire classes of heap
overflow exploits and similar kernel memory exposures.

Change-Id: I17cf6c6296d63659e52d29e2c41f105843e2960e
Signed-off-by: Shivendra Pratap <quic_spratap@quicinc.com>