PENDING: arm64: dts: Increase min voltage of l1b and l10c rails
Some of the clients of l1b and l10c regulators do not vote for voltage.
The min voltage requirement for these clients is more than what is
currently specified.

Add change to increase the min voltage of these rails to a level
which is sufficient for it's client requirements.

While at it, revert commit 6c4538162585 ("PENDING: arm64:
dts: qcm6490: Remove voltage voting for USB rails").

Change-Id: I4a7acfabe2a4200f90d4bf6459dae7548d54f248
Signed-off-by: Umang Chheda <quic_uchheda@quicinc.com>