PENDING: media: qcom: camss: Attach formats to VFE resources
Video node formats have direct dependency by the sub-device pad formats.
Remove dependency for SoC version and move format definitions in device
which creates video node.

This commit attaches a struct to the VFE resources that holds format
description, so it is much easier to assign them to the video node.
No need to use a switch-case.

NOTE: The mbus_bpp is used to calculate the clock rates and is different
from bpp which is the actual bpp written to memory. We need to keep
both values to not break the calculations.

Change-Id: If2be3b2d91932721abee7de8cfa493cbb535ffc5
Signed-off-by: Radoslav Tsvetkov <quic_rtsvetko@quicinc.com>
Signed-off-by: Vikram Sharma <quic_vikramsa@quicinc.com>