FROMLIST: drm/drm_mode_object: increase max objects to accommodate new color props
DRM_OBJECT_MAX_PROPERTY limits the number of properties to be attached
and we are increasing that value all time we add a new property (generic
or driver-specific).

In this series, we are adding 13 new KMS driver-specific properties for
AMD color manage:
- CRTC Gamma enumerated Transfer Function
- Plane: Degamma LUT+size+TF, HDR multiplier, shaper LUT+size+TF, 3D
  LUT+size, blend LUT+size+TF (12)

Therefore, just increase DRM_OBJECT_MAX_PROPERTY to a number (64) that
accommodates these new properties and gives some room for others,
avoiding change this number every time we add a new KMS property.

Change-Id: I990a69bd5b423fe8433daf89ead4358a6c5dce30
Reviewed-by: Harry Wentland <harry.wentland@amd.com>
Reviewed-by: Simon Ser <contact@emersion.fr>
Signed-off-by: Melissa Wen <mwen@igalia.com>
Link: https://lore.kernel.org/all/20231116195812.906115-2-mwen@igalia.com/.
Patch-mainline: dri-devel @ 11/16/23, 19:58
Signed-off-by: Mahadevan <quic_mahap@quicinc.com>