PENDING: misc: fastrpc: Fix DSP capabilities request
Incorrect remote arguments are getting passed when requesting for
capabilities from DSP. Also there is no requirement to update the
PD type as it might cause problems for any PD other than user PD.
In addition to this, the collected capability information is not
getting copied properly to user. Add changes to address these
problems and get correct DSP capabilities.

Change-Id: Id6ac44e07e3aca4780936cb55b8f0efec6fce16a
Signed-off-by: Ekansh Gupta <quic_ekangupt@quicinc.com>