PENDING: misc: fastrpc: Add static PD restart support
Static PDs on the audio and sensor domains are expected to support
PD restart. The kernel resource handling for the PDs are expected
to be handled by fastrpc driver. For this, there is a requirement
of PD service locator to get the event notifications for static PD
services. Also when events are received, the driver needs to handle
based on PD states. Added changes to add service locator for audio
and sensor domain static PDs and handle the PD restart sequence.

Change-Id: I73bb1ccfe70a2ef938a9206ab1f35c5800d6217c
Signed-off-by: Ekansh Gupta <quic_ekangupt@quicinc.com>