FROMLIST: iommu/arm-smmu: introduction of ACTLR for custom prefetcher settings
Currently in Qualcomm  SoCs the default prefetch is set to 1 which allows
the TLB to fetch just the next page table. MMU-500 features ACTLR
register which is implementation defined and is used for Qualcomm SoCs
to have a prefetch setting of 1/3/7/15 enabling TLB to prefetch
the next set of page tables accordingly allowing for faster translations.

ACTLR value is unique for each SMR (Stream matching register) and stored
in a pre-populated table. This value is set to the register during
context bank initialisation.

Signed-off-by: Bibek Kumar Patro <quic_bibekkum@quicinc.com>

Change-Id: I3df452ea8b2cf0fdefbddb2c037a4d228a0f318a
Link: https://lore.kernel.org/all/20231127145412.3981-2-quic_bibekkum@quicinc.com/
Patch-mainline: linux-arm-msm @ 2023-11-27 14:54
[quic_bibekkum@quicinc.com: updated the copyright year to 2023]
Signed-off-by: Bibek Kumar Patro <quic_bibekkum@quicinc.com>