PENDING: arm64: dts: qcm6490-rb3: Increase min voltage of l1b & l10c rail
Some of the clients of l1b and l10c regulators do not vote for voltage.
The min voltage requirement for these clients is more than what is
currently specified.

Add change to increase the min voltage of these rails to a level
which is sufficient for it's client requirements.

While at it, revert commit bf0990d9ce67 ("PENDING: arm64: dts:
qcm6490-rb3: Remove voltage voting for USB rails").

Change-Id: If448ad78ea323cd811bae6feb79bc33750e44ef2
Signed-off-by: Umang Chheda <quic_uchheda@quicinc.com>