PENDING: usb: dwc3: qcom: De-assert BCR upon resume
Since the rework of qualcomm dwc3 wrapper includes flattening the
device tree node and using core as a library instead of device, the
BCR reset can be handled either in core or in wrapper. The rework
leaves the assertion/de-assertion of BCR to core driver.

Once the device is suspended, core_exit would assert the BCR but
during resume, wrapper's resume routine runs and tries to enable
clocks and interconnects. But at this instant the BCR is still
asserted and hence the clocks would stay stuck at OFF resulting in
a crash.

De-assert the BCR upon wrapper's resume to avoid the crash.

Change-Id: I59ba0fbb93a09bc100a0beb80096566f7ca7d061
Signed-off-by: Krishna Kurapati <quic_kriskura@quicinc.com>