PENDING: misc: fastrpc: Release PDR handle
PDR handle is allocated during rpmsg_probe functionality which
is used to add lookup for PDR notifications. This handle is
not released which might cause problems for SSR scenario as
the probe functionality will again get called which will
try to allocate the handle again. Add changes to release the
allocated handle during rpmsg_remove call.

Change-Id: I3e73f544d80120e9d7b13e46e00ff85ae09caf94
Signed-off-by: Ekansh Gupta <quic_ekangupt@quicinc.com>