PENDING: iris: change the default value of HEADER_MODE config
The HEADER_MODE control decides whether to join the codec
config data and first I frame data into a single buffer
for encoder usecase. Currently, the default value of this
config ensures this data to be sent separately. This change
modifies the default value to send codec config and I frame
data in a single buffer.

Change-Id: I5e78a97069a768b66c53afe0d81c20a0d1fb8cbd
Signed-off-by: Chandrakant I Viraktamath <quic_civirakt@quicinc.com>