PENDING: arm64: boot: dts: Add Support for nb7vpq904m on QCM6490 RB3
QCM6490 RB3 has a re-driver sitting in the board responsible to detecting
orientation info and boost RX/TX signalling. It established connection with
PMIC with I2C procotol and get info to USB for orientation detection.

This change enabled the dt node linked to usb and pmic for the
communication.

Change-Id: Id4b69d713fe0a45bedb9a55d11de767e01b5e58c
Signed-off-by: Prashanth K <quic_prashk@quicinc.com>
Signed-off-by: Udipto Goswami <quic_ugoswami@quicinc.com>