PENDING: iris: add support for dynamic resolution flag
Add support for V4L2_FMT_FLAG_DYN_RESOLUTION flag
to the coded formats of the decoder returned
during enumerate format IOCTL calls.

This flag indicates that the dynamic resolution
switching is supported by the device for the
compressed format. It will notify the user via the event
V4L2_EVENT_SOURCE_CHANGE when changes in the video
parameters are detected.

Change-Id: I27cc6419c7eb7035f578c82a118ddc63a16445b8
Signed-off-by: Chandrakant I Viraktamath <quic_civirakt@quicinc.com>