PENDING: firmware: shmbridge: Defer probe if qcom_scm is not up
If loading qcom_scm fails for some reason and as shmbridge is dependent
on qcom_scm, defer shmbridge probe sequence till qcom_scm is available.

Alongwith this, also check if memory pool is available or not before
trying to allocate memory from the pool. Without this check, we may
observe failures if shmbridge is not loaded successfully and client
tries to allocate memory via shmbridge pool.

Change-Id: I132d766ce6426f03c8952f597e98b0d2c8ee19e1
Signed-off-by: Kuldeep Singh <quic_kuldsing@quicinc.com>