FROMLIST: coresight: core: Add device name support
For some coresight components like CTI and TPDM, there could be
numerous of them. From the node name, we can only get the type and
register address of the component. We can't identify the HW or the
system the component belongs to. Add the device-name support for
adding the intuitive name of the device.

Link: https://patchwork.kernel.org/project/linux-arm-msm/patch/20240131082628.6288-2-quic_jinlmao@quicinc.com/
Change-Id: I285842eed4e7b50b7ed3c93b1e0ba7f43aa2cecc
Signed-off-by: Mao Jinlong <quic_jinlmao@quicinc.com>