FROMLIST: dt-bindings: arm: Add device-name in the coresight components
Current name of coresight component's folder consists of prefix of
the device and the id in the device list. When run 'ls' command,
we can get the register address of the device. Take CTI for example,
if we want to set the config for modem CTI, but we can't know which
CTI is modem CTI from all current information.

cti_sys0 -> ../../../devices/platform/soc@0/138f0000.cti/cti_sys0
cti_sys1 -> ../../../devices/platform/soc@0/13900000.cti/cti_sys1

Add device-name in device tree which can provide a better description
of the coresight device. It can provide the info like the system or
HW it belongs to.

Link: https://patchwork.kernel.org/project/linux-arm-msm/patch/20240131082628.6288-3-quic_jinlmao@quicinc.com/
Change-Id: I07d9e53f8454d2de00e02c8182e814d91bdec3e2
Signed-off-by: Mao Jinlong <quic_jinlmao@quicinc.com>