PENDING: interconnect: qcom: enable sync_state for SM8550
The bootloaders often do some initial configuration of the interconnects
in the system and we want to keep this configuration until all consumers
have probed and expressed their bandwidth needs. This is because we don't
want to change the configuration by starting to disable unused paths until
every user had a chance to request the amount of bandwidth it needs.

Change-Id: I05910b09d9007280143c882f46d1546964169d86
Signed-off-by: Raviteja Laggyshetty <quic_rlaggysh@quicinc.com>