spi: spi-msm-geni: Enable or disable the SSC clock in LPM path
If we get the system suspend request after ssr up sequence then
in LPM path we are not turning off the core/core2x clock and due
to this XO shutdown test case will fail. To avoid this issue
enable or disable the core/core2x clock in suspend/resume path
based on the enable flag.

Change-Id: I82778341fae51c2c0d37eaa6d4728e823393c4e3
Signed-off-by: Aniket Randive <quic_arandive@quicinc.com>
Signed-off-by: Zahir Shabbir Khan <quic_zshabbir@quicinc.com>