QCLINUX: linux: soc: qcom: Fix usb_qdss_alloc_req with valid parameters
Currently, the definition of usb_qdss_alloc_req is wrong is case the
defconfig is not enabled as the function take 3 argument but
there as only 2 of them passed by the qdss client driver. This
gives out a compilation failure is cases where CONFIG_USB_F_QDSS is
disabled.

Fix this by removing the invalid extra function parameter. Also
address NULL pointer in the channel name due to wrong channel
variable usage.

Change-Id: I8b3dc444a59c6fa24f3162577c80c503cef3a4ba
Signed-off-by: Udipto Goswami <quic_ugoswami@quicinc.com>